Defining the structure that store the details of a dog

+++ structures_typedef/dog.h
@@ -4,7 +4,7 @@
 /**
  * struct dog - collects personal information
  * @name: stores string of charactors as name of dog
- * @age: stores integers as age of dog
+ * @age: stores a float numeral as age of dog
  * @owner: array of charactors as name of owner
  *
  * Description: this structure stores the details of a dog
@@ -13,6 +13,6 @@
 struct dog 
 {
 	char *name, *owner;
-	int age;
+	float age;
 };
 #endif
